+++ NEWS.md
@@ -2,6 +2,7 @@
 
 ## Next
 
+* New function `company-grab-symbol-cons`.
 * `company-clang` fetches completion candidates asynchronously.
 * Added support for asynchronous back-ends (experimental).
 * Support for back-end command `crop` dropped (it was never documented).